Refactor import statements in page.tsx and Header.tsx
