+++ frontend/app/common/components/index.ts
@@ -1,0 +1,3 @@
+export * from './MobileCategories'
+export * from './MobileMenu'
+export * from './MobileSearch'
